requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic-settings
supabase
python-jose[cryptography]
pytest
pytest-asyncio
ruff
openai==1.82.1
python-multipart
